move FileRef and BundleFileType to bundle.py
