+++ reconcile/test/change_owners/fixtures.py
@@ -9,10 +9,12 @@ import jsonpath_ng
 import jsonpath_ng.ext
 import pytest
 
-from reconcile.change_owners.change_types import (
-    BundleFileChange,
+from reconcile.change_owners.bundle import (
     BundleFileType,
     FileRef,
+)
+from reconcile.change_owners.change_types import (
+    BundleFileChange,
     create_bundle_file_change,
 )
 from reconcile.gql_definitions.change_owners.queries import self_service_roles
