+++ reconcile/test/change_owners/test_change_type_diff_splitting.py
@@ -1,8 +1,10 @@
-from reconcile.change_owners.change_types import (
+from reconcile.change_owners.bundle import (
     BundleFileType,
+    FileRef,
+)
+from reconcile.change_owners.change_types import (
     ChangeTypeContext,
     ChangeTypeProcessor,
-    FileRef,
     build_change_type_processor,
     create_bundle_file_change,
 )
